+++ src/aiheuristics/EmptySpaces.java
@@ -27,11 +27,11 @@ public class EmptySpaces implements Heuristic {
 
     @Override
     public long getValueOfState(GameController controller, GameBoard state, int currentDirection) {
-        int[][] grid = state.getGameGrid();
+        byte[][] grid = state.getGameGrid();
         int[] highestPos = getHighestValuePosition(grid);
         int highestValue = grid[highestPos[0]][highestPos[1]];
         long output = 0;
-        for (int[] row : grid) {
+        for (byte[] row : grid) {
             for (int value : row) {
                 if (value == 0) {
                     output += (highestValue / 2);
@@ -43,11 +43,11 @@ public class EmptySpaces implements Heuristic {
     }
 
 // returned as y, x
-    private int[] getHighestValuePosition(int[][] grid) {
+    private int[] getHighestValuePosition(byte[][] grid) {
         int[] highest = new int[2];
         int highestValue = 0;
         for (int y = 0; y < grid.length; y++) {
-            int[] row = grid[y];
+            byte[] row = grid[y];
             for (int x = 0; x < row.length; x++) {
                 int value = row[x];
                 if (value > highestValue) {
